// schematics/src/page/schema.d.ts
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * Creates a new page.
 */
export interface PwaPageOptionsSchema {
  path?: string;
  restricted?: boolean;
  tsext?: string;
  project?: string;
  /**
   * The name of the page.
   */
  name?: string;
  /**
   * The extension to generate the page in. If empty it is generated in the project root.
   */
  extension?: string;
}